Update PhysX to 5.6.1 at b07ddd12

+++ physx/include/foundation/unix/PxUnixMathIntrinsics.h
@@ -142,7 +142,7 @@ PX_CUDA_CALLABLE PX_FORCE_INLINE bool isFinite(double a)
 /*!
 Sets \c count bytes starting at \c dst to zero.
 */
-PX_FORCE_INLINE void* memZero(void* dest, uint32_t count)
+PX_FORCE_INLINE void* memZero(void* dest, size_t count)
 {
 	return memset(dest, 0, count);
 }
@@ -150,7 +150,7 @@ PX_FORCE_INLINE void* memZero(void* dest, uint32_t count)
 /*!
 Sets \c count bytes starting at \c dst to \c c.
 */
-PX_FORCE_INLINE void* memSet(void* dest, int32_t c, uint32_t count)
+PX_FORCE_INLINE void* memSet(void* dest, int32_t c, size_t count)
 {
 	return memset(dest, c, count);
 }
@@ -158,7 +158,7 @@ PX_FORCE_INLINE void* memSet(void* dest, int32_t c, uint32_t count)
 /*!
 Copies \c count bytes from \c src to \c dst. User memMove if regions overlap.
 */
-PX_FORCE_INLINE void* memCopy(void* dest, const void* src, uint32_t count)
+PX_FORCE_INLINE void* memCopy(void* dest, const void* src, size_t count)
 {
 	return memcpy(dest, src, count);
 }
@@ -166,7 +166,7 @@ PX_FORCE_INLINE void* memCopy(void* dest, const void* src, uint32_t count)
 /*!
 Copies \c count bytes from \c src to \c dst. Supports overlapping regions.
 */
-PX_FORCE_INLINE void* memMove(void* dest, const void* src, uint32_t count)
+PX_FORCE_INLINE void* memMove(void* dest, const void* src, size_t count)
 {
 	return memmove(dest, src, count);
 }
